seperate btc and eth link tests
