+++ pallets/account-linker/src/tests/mod.rs
@@ -1,0 +1,2 @@
+mod btc;
+mod eth;
